funtion for number of element

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,14 +1,15 @@
 #include "lists.h"
 
 /**
- * print_listint - function prints all list elements
- * @h: head of list
+ * print_listint - prints all list elements.
+ * @h: head of the list.
  *
- * Return: number of nodes
+ * Return: number of nodes.
  */
 size_t print_listint(const listint_t *h)
 {
-	int i = 0;
+	size_t i = 0;
+
 	while (h != NULL)
 	{
 		printf("%d\n", h->n);
